API Resource for Project

+++ bdcrrm_server/services/project/service.py
@@ -13,6 +13,7 @@ from typing import Dict, List
 import werkzeug.exceptions as werkzeug_exceptions
 from invenio_records_resources.services import Service
 
+from ...forms import ProjectForm
 from ...models import Project, ProjectUser
 from ...models import db
 
@@ -20,11 +21,11 @@ from ...models import db
 class ProjectService(Service):
     """Project Service."""
 
-    def create_project(self, user_id, data) -> Dict:
+    def create_project(self, identity, data) -> Dict:
         """Create a Project.
 
         Args:
-            user_id (int): Project Owner User ID (from OAuth service)
+            identity (flask_principal.Identity): Project Owner User ID (from OAuth service)
 
             data (dict): Project data received from the user.
         Returns:
@@ -32,6 +33,12 @@ class ProjectService(Service):
         Raises:
             Exception: When Project is not created.
         """
+        # validating
+        form = ProjectForm()
+        form.load(data)
+
+        data["_metadata"] = data["metadata"]
+
         with db.session.begin_nested():
             created_project = Project(**data)
 
@@ -40,7 +47,7 @@ class ProjectService(Service):
 
             project_user = ProjectUser(
                 project_id=created_project.id,
-                user_id=user_id,
+                user_id=identity.id,
                 is_admin=True,  # 'Owner' is admin
                 active=True
             )
@@ -48,23 +55,23 @@ class ProjectService(Service):
         db.session.commit()
         return created_project
 
-    def list_project_by_user(self, user_id: int) -> List[Project]:
+    def list_project_by_user(self, identity) -> List[Project]:
         """List Project by a specific user
 
         Args:
-            user_id (int): Project User ID (from OAuth service)
+            identity (flask_principal.Identity): Project User identity (from OAuth service)
         Returns:
             List[Project]: List of `user_id` projects.
         """
         return db.session.query(
             Project
-        ).filter(ProjectUser.user_id == user_id).all()
+        ).filter(ProjectUser.user_id == identity.id).all()
 
-    def get_project_by_id(self, user_id: int, project_id: int) -> Project:
+    def get_project_by_id(self, identity, project_id: int) -> Project:
         """Get a project by `user` and `project` id.
 
         Args:
-            user_id (int): Project User ID (from OAuth service)
+            identity (flask_principal.Identity): Project User identity (from OAuth service)
 
             project_id (int): Project ID
         Returns:
@@ -72,16 +79,16 @@ class ProjectService(Service):
         """
         user_projects = db.session.query(ProjectUser).filter(
             ProjectUser.project_id == project_id,
-            ProjectUser.user_id == user_id
+            ProjectUser.user_id == identity.id
         ).first_or_404("Project not found!")
 
         return user_projects.project
 
-    def delete_project_by_id(self, user_id: int, project_id: int):
+    def delete_project_by_id(self, identity, project_id: int):
         """Delete a project object on database.
 
         Args:
-            user_id (int): Project User ID (from OAuth service)
+            identity (flask_principal.Identity): Project User identity (from OAuth service)
 
             project_id (int): Project ID that will be deleted.
         Returns:
@@ -89,23 +96,25 @@ class ProjectService(Service):
         """
         selected_user = db.session.query(ProjectUser).filter(
             ProjectUser.project_id == project_id,
-            ProjectUser.user_id == user_id
+            ProjectUser.user_id == identity.id
         ).first_or_404("Project not found!")
 
         if not selected_user.is_admin:
             raise werkzeug_exceptions.Unauthorized(description="Admin access is required to delete the project.")
 
         with db.session.begin_nested():
-            db.session.delete(selected_user.project.graph)
+            if selected_user.project.graph:
+                db.session.delete(selected_user.project.graph)
+
             db.session.delete(selected_user.project)
             db.session.delete(selected_user)
         db.session.commit()
 
-    def edit_project_by_id(self, project_id, user_id, attributes_to_chage: Dict) -> Project:
+    def edit_project_by_id(self, identity, project_id, attributes_to_chage: Dict) -> Project:
         """Edit a project object on database.
 
         Args:
-            user_id (int): Project User ID (from OAuth service)
+            identity (flask_principal.Identity): Project User identity (from OAuth service)
 
             project_id (int): Project ID that will be deleted.
 
@@ -113,9 +122,13 @@ class ProjectService(Service):
         Returns:
             Project: The project updated on the database.
         """
+        # validating
+        form = ProjectForm()
+        form.load(attributes_to_chage, partial=True)
+
         selected_user = db.session.query(ProjectUser).filter(
             ProjectUser.project_id == project_id,
-            ProjectUser.user_id == user_id
+            ProjectUser.user_id == identity.id
         ).first_or_404("Project not found!")
 
         if not selected_user.is_admin:
